requirements.txt
tensorflow
mss
numpy
opencv-python
Pillow

